LayoutTests/media/modern-media-controls/start-support/start-support-click-to-start.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
        width: 800px;
        height: 600px;
    }
    
</style>
<video src="../../content/test.mp4" controls></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>MediaController</code> click-to-start behavior.");

const host = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(host, media, null);

media.addEventListener("loadedmetadata", function() {
    shouldBeTrue("mediaController.controls.showsStartButton");
    debug("");
    debug("Pressing on the start button");
    shouldBecomeDifferent("mediaController.controls.playPauseButton.element.getBoundingClientRect().width", "0", () => {
        pressOnElement(mediaController.controls.playPauseButton.element)
    });
});

media.addEventListener("play", function() {
    debug("Media is playing");
    shouldBeFalse("mediaController.controls.showsStartButton");
    debug("");
    host.remove();
    media.remove();
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
